+++ CyanUtil/src/main/java/org/asf/cyan/api/packet/entries/ByteEntry.java
@@ -24,7 +24,7 @@ public class ByteEntry implements PacketEntry<Byte> {
 	}
 
 	@Override
-	public boolean isCompatible(long type) {
+	public boolean isCompatible(byte type) {
 		return type == type();
 	}
