+++ CyanUtil/src/main/java/org/asf/cyan/api/packet/entries/LongEntry.java
@@ -25,7 +25,7 @@ public class LongEntry implements PacketEntry<Long> {
 	}
 
 	@Override
-	public boolean isCompatible(long type) {
+	public boolean isCompatible(byte type) {
 		return type == type();
 	}
